+++ src/metricsAdapter.js
@@ -1,5 +1,12 @@
 var Metrics = require( 'metrics' );
 var report = new Metrics.Report();
+require( './index' );
+var convert = require( './converter' );
+
+var types = {
+	'time': 'Timer',
+	'meter': 'Histogram'
+};
 
 function createMetric( type, name ) {
 	var metric = report.getMetric( name );
@@ -10,22 +17,24 @@ function createMetric( type, name ) {
 	return metric;
 }
 
-function getReport() {
-	return report.summary();
+function getReport( reset ) {
+	var data = report.summary();
+	if ( reset ) {
+		report = new Metrics.Report();
+	}
+	return data;
 }
 
-function recordDuration( name, duration ) {
-	var timer = createMetric( 'Timer', name );
-	timer.update( duration );
-}
-
-function recordMeter( name, value ) {
-	var meter = createMetric( 'Histogram', name );
-	meter.update( value );
+function recordMetric( data ) {
+	var value = data.value;
+	if ( data.type === 'time' ) {
+		value = convert( data.value, data.units, 'ms' );
+	}
+	var metric = createMetric( types[ data.type ], data.key );
+	metric.update( value );
 }
 
 module.exports = {
 	getReport: getReport,
-	onMeter: recordMeter,
-	onTime: recordDuration
+	onMetric: recordMetric
 };